#184: Rename message.js to chat.js
